+++ crawler/src/test/java/in/nimbo/isDoing/searchEngine/crawler/lru/CaffeineLRUTest.java
@@ -9,13 +9,13 @@ import org.junit.Before;
 import org.junit.Test;
 
 import java.io.FileInputStream;
-import java.net.URL;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 public class CaffeineLRUTest {
