+++ crawler/src/test/java/in/nimbo/isDoing/searchEngine/crawler/lru/LRULinkHashMapImplTest.java
@@ -5,10 +5,10 @@ import in.nimbo.isDoing.searchEngine.engine.Status;
 import in.nimbo.isDoing.searchEngine.pipeline.Output;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 public class LRULinkHashMapImplTest {
     @Before
