+++ engine/pom.xml
@@ -36,8 +36,9 @@
         <dependency>
             <groupId>org.apache.hbase</groupId>
             <artifactId>hbase-client</artifactId>
-            <version>2.1.0</version>A
+            <version>2.1.0</version>
         </dependency>
+
         <dependency>
             <groupId>com.joyent.util</groupId>
             <artifactId>fast-md5</artifactId>
