+++ engine/src/main/resources/log4j.properties
@@ -5,7 +5,7 @@ log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
 log4j.appender.stdout.layout.ConversionPattern=%-4r [%t] %-5p %c %x - %m%n
 
 log4j.appender.kafkaAppender=org.apache.log4j.FileAppender
-log4j.appender.kafkaAppender.File=in.nimbo.isDoing.searchEngine.kafka.log
+log4j.appender.kafkaAppender.File=kafka.log
 log4j.appender.kafkaAppender.layout=org.apache.log4j.PatternLayout
 log4j.appender.kafkaAppender.layout.ConversionPattern= %-4r [%t] %-5p %c %x - %m%n
 
@@ -15,7 +15,7 @@ log4j.appender.counter.layout=org.apache.log4j.PatternLayout
 log4j.appender.counter.layout.ConversionPattern= %m%n
 
 log4j.appender.fileAppender=org.apache.log4j.FileAppender
-log4j.appender.fileAppender.File=search-in.nimbo.isDoing.searchEngine.engine.log
+log4j.appender.fileAppender.File=search-engine.log
 log4j.appender.fileAppender.layout=org.apache.log4j.PatternLayout
 log4j.appender.fileAppender.layout.ConversionPattern= %-4r [%t] %-5p %c %x - %m%n
